test (Frame) : 리팩토링하면서 NormalFrame, FinalFrame test 코드 변경
